fix for the long-standing problem with not being able to do higher baudrates than 19200... this may even work on revA. Also fixed the new TWI configuration options, which seems to work (at least for the ADXL345, the others will be tested soon).

+++ MiniBee/MiniBee.h
@@ -81,7 +81,7 @@ class MiniBee {
 		
 // 		void ParseCustom( char * msg ){ Serial.println( msg ); };
 		
-		void begin(int); //init function
+		void begin(long); //init function
 		void doLoopStep(void); // loop function
 		
 		void setRemoteConfig( bool onoff );
@@ -98,7 +98,7 @@ class MiniBee {
 
 		void readXBeeSerial(void);
 		
-		void openSerial(int);
+		void openSerial(long);
 		void configXBee();
 
 		void setID( uint8_t id );
